Extract function: loading path objects from segmasks
