+++ src/main/java/Images.kt
@@ -1,9 +1,17 @@
+import Images.Companion.logger
+import ij.IJ
+import ij.process.ColorProcessor
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
+import qupath.imagej.processing.RoiLabeling
+import qupath.imagej.tools.IJTools
 import qupath.lib.gui.commands.ProjectCommands
 import qupath.lib.images.ImageData
 import qupath.lib.images.servers.ImageServerProvider
+import qupath.lib.objects.PathObject
+import qupath.lib.objects.PathObjects
 import qupath.lib.projects.Project
+import qupath.lib.regions.ImagePlane
 import java.awt.image.BufferedImage
 import java.io.File
 
@@ -46,4 +54,35 @@ fun Project<BufferedImage>.addImages(inputImages: List<InputImage>) {
     }
 
   this.syncChanges()
+}
+
+fun extractPathObjects(
+  maskPath: String,
+  downsample: Double = 1.0,
+  imagePlane: ImagePlane = ImagePlane.getDefaultPlane(),
+): List<PathObject> {
+  val imp = IJ.openImage(maskPath)
+  logger.info(imp.toString())
+  val n = imp.statistics.max.toInt()
+  logger.info("   Max Cell Label: $n")
+  if (n == 0) {
+    logger.info(" >>> No objects found! <<<")
+    return emptyList()
+  }
+
+  val ip = imp.processor
+  if (ip is ColorProcessor) {
+    throw IllegalArgumentException("RGB images are not supported!")
+  }
+
+  val roisIJ = RoiLabeling.labelsToConnectedROIs(ip, n)
+  val rois = roisIJ.mapNotNull {
+    if (it == null) {
+      null
+    } else {
+      IJTools.convertToROI(it, 0.0, 0.0, downsample, imagePlane)
+    }
+  }
+
+  return rois.map { PathObjects.createDetectionObject(it) }
 }
